Move to cmd directory
