--- hack/main.go
@@ -1,14 +1,0 @@
-package main
-
-import (
-	"os"
-)
-
-func main() {
-	switch os.Args[1] {
-	case "removecrdvalidation":
-		removeCRDValidation(os.Args[2])
-	default:
-		panic(os.Args[1])
-	}
-}
