--- hack/types.go
@@ -1,3 +1,0 @@
-package main
-
-type obj = map[string]interface{}
